Mean execution time, instead of total time

+++ test.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
+# check if argument is either FCFS or SJF
+if [ "$1" != "FCFS" ] && [ "$1" != "SJF" ]; then
+    echo "Usage: $0 <FCFS|SJF>"
+    exit 1
+fi
+
 rm -f tmp/* /tmp/escalonador*
-time bin/orchestrator tmp 3 FCFS && ./mean_time.sh tmp/completed.txt &
+time bin/orchestrator tmp 3 $1 && ./mean_time.sh tmp/completed.txt && echo '' &
 
 bin/client execute 100 -u 'test/hello 6' > /dev/null
 bin/client execute 100 -u 'test/hello 6' > /dev/null
